+++ packages/editor/web/src/RichContentEditor/RichContentEditor.jsx
@@ -97,7 +97,7 @@ class RichContentEditor extends Component {
 
   deprecateSiteDomain = () => {
     const { config, siteDomain } = this.props;
-    if (config[HTML_TYPE]) {
+    if (config[HTML_TYPE] && siteDomain) {
       config[HTML_TYPE].siteDomain = siteDomain;
     }
   };