+++ packages/plugin-html/web/src/icons/index.js
@@ -1,2 +1,3 @@
 export { default as EditIcon } from './EditIcon';
 export { default as InsertPluginIcon } from './InsertPluginIcon';
+export { default as AdsenseIcon } from './Adsense.svg';