+++ packages/plugin-link-preview/web/src/toolbar/insert-buttons.jsx
@@ -46,7 +46,7 @@ export default ({ helpers, settings, isMobile, t }) => {
       helpers,
       modalElement: EmbedURLInputModal,
       modalStyles: getModalStyles({ customStyles, fullScreen: false, isMobile }),
-      section: 'BlockToolbar_Section_Embed_Social',
+      section: 'BlockToolbar_Section_Embed_Anywhere',
     };
   });
 };